atc: structure: use vars.Any for var source config
Doing so lets us avoid the awkwardness of needing to wrap the config in
an `atc.Params` with a throwaway key in order to interpolate it.

In principle, this change could let you specify the full var source
config as a variable:

name: whatever
type: vault
config: ((full_config))

...but, the way we validate var_sources currently makes that impossible.
It also makes it impossible to interpolate certain fields in the config.

I also realized we didn't have any tests that tested the happy path for
a var_source in configvalidate, so I added that

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>