atc: behaviour: quote slash in instance var value
Technically, you don't need to quote the slash when you specify a
pipeline ref in fly via `-j 'release/branch:feature/foo/job'` - fly
knows to take everything between the first / and the last / and treat
that as the instance vars. However, it'll probably make it easier for
other tools to parse pipeline refs if we quote the / when we generate
it.

Also, I think `-j 'release/branch:"feature/foo"/job'` is a bit nicer to
read.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>