atc: structure: use vars types for Source
I initially did the same for VersionedResourceTypes (directly running
vars.InterpolateInto(...)), but this changed the behaviour by
interpolating Params as well. We thankfully had a test to check that
Params was left uninterpolated.

So, unfortunately, we can't get rid of this creds.xxx helper - but I
simplified it a little by removing the intermediate type.

Signed-off-by: Aidan Oldershaw <aoldershaw@pivotal.io>